Split client_server to client.js and server.js
