+++ test/client.js
@@ -1,12 +1,10 @@
 'use strict';
 
 const accounts = require('../index.js');
-const cache = require('../src/cache');
 
 const chai = require('chai');
 const chai_as_promised = require('chai-as-promised');
 const nock = require('nock');
-const httpMocks = require('node-mocks-http');
 
 
 chai.should();
@@ -36,7 +34,7 @@ const payloads = {
     }
 };
 
-describe('accounts client', () => {
+describe('generate_token', () => {
 
     accounts.configure(configuration);
 
@@ -56,7 +54,6 @@ describe('accounts client', () => {
             });
     });
 
-
     it('should get expired access token again with specified scope', () => {
         let accounts_nock = nock(configuration.base_url)
             .post(configuration.path + '/oauth/token', payloads.client_credentials)
@@ -72,7 +69,6 @@ describe('accounts client', () => {
                 accounts_nock.isDone().should.equal(true);
             });
     });
-
 
     it('should get cached access token with specified scope', () => {
         let accounts_nock = nock(configuration.base_url)
@@ -109,7 +105,7 @@ describe('accounts client', () => {
     });
 });
 
-describe('refresh token', () => {
+describe('refresh_token', () => {
 
     accounts.configure(configuration);
 
@@ -150,7 +146,7 @@ describe('refresh token', () => {
             });
     });
 
-    it('should properly updated cache of refreshed token', () => {
+    it('should properly update cache of refreshed token', () => {
         let accounts_nock = nock(configuration.base_url)
             .post(configuration.path + '/oauth/token', payloads.client_credentials)
             .reply(200, {access_token: 'jrrtoken', refresh_token: 'jrrrefreshtoken'});
@@ -166,163 +162,6 @@ describe('refresh token', () => {
                 accounts_nock.isDone().should.not.equal(true);
 
                 nock.cleanAll();
-            })
-    });
-});
-
-describe('verify scopes', () => {
-    it('should fail without a provided access token', () => {
-        const request = httpMocks.createRequest({
-            method: 'GET',
-            url: '/test',
-            params: {}
-        });
-
-        return accounts.verify_scopes(scopes)(
-            request,
-            null,
-            error => {
-                return error ? Promise.reject(error) : Promise.resolve();
-            }
-        ).catch(error => {
-            error.should.have.property('message');
-            error.message.should.be.a('string');
-            error.message.should.equal('Access token is required.');
-        });
-    });
-
-    it('should forward remote failure', () => {
-        let accounts_nock = nock(configuration.base_url)
-            .get(configuration.path + '/oauth/tokeninfo')
-            .query({access_token: 'notjrrtoken'})
-            .reply(500, {message: 'random server error'});
-
-        const request = httpMocks.createRequest({
-            method: 'GET',
-            url: '/test',
-            headers: {'Access-Token': 'notjrrtoken'}
-        });
-
-
-        return accounts.verify_scopes(scopes)(
-            request,
-            null,
-            error => {
-                return error ? Promise.reject(error) : Promise.resolve();
-            }
-        ).catch(error => {
-            error.should.have.property('message');
-            error.message.should.be.a('string');
-            error.message.should.equal('random server error');
-            accounts_nock.isDone().should.equal(true);
-        });
-    });
-
-    it('should fail without returned scopes', () => {
-        let accounts_nock = nock(configuration.base_url)
-            .get(configuration.path + '/oauth/tokeninfo')
-            .query({access_token: 'notjrrtoken'})
-            .reply(200, {});
-
-        const request = httpMocks.createRequest({
-            method: 'GET',
-            url: '/test',
-            headers: {'Access-Token': 'notjrrtoken'}
-        });
-
-
-        return accounts.verify_scopes(scopes)(
-            request,
-            null,
-            error => {
-                return error ? Promise.reject(error) : Promise.resolve();
-            }
-        ).catch(error => {
-            error.should.have.property('message');
-            error.message.should.be.a('string');
-            error.message.should.equal('Something went wrong, server did not return scopes, please try again.');
-            accounts_nock.isDone().should.equal(true);
-        });
-    });
-
-    it('should fail without any of the required scopes', () => {
-        let accounts_nock = nock(configuration.base_url)
-            .get(configuration.path + '/oauth/tokeninfo')
-            .query({access_token: 'notjrrtoken'})
-            .reply(200, {scopes: 'http://non.existing/scope'});
-
-        const request = httpMocks.createRequest({
-            method: 'GET',
-            url: '/test',
-            headers: {'Access-Token': 'notjrrtoken'}
-        });
-
-
-        return accounts.verify_scopes(scopes)(
-            request,
-            null,
-            error => {
-                return error ? Promise.reject(error) : Promise.resolve();
-            }
-        ).catch(error => {
-            error.should.have.property('message');
-            error.message.should.be.a('string');
-            error.message.should.equal(
-                'You need at least one of the ff. scopes to access this endpoint: ' + scopes.join(' ')
-            );
-            accounts_nock.isDone().should.equal(true);
-        });
-    });
-
-    it('should succeed with proper provided access token', () => {
-        let accounts_nock = nock(configuration.base_url)
-            .get(configuration.path + '/oauth/tokeninfo')
-            .query({access_token: 'jrrtoken'})
-            .reply(200, {scopes: scopes.join(' ')});
-
-        const request = httpMocks.createRequest({
-            method: 'GET',
-            url: '/test',
-            headers: {'Access-Token': 'jrrtoken'}
-        });
-
-
-        return accounts.verify_scopes(scopes)(
-            request,
-            null,
-            error => {
-                return error ? Promise.reject(error) : Promise.resolve();
-            }
-        ).catch(error => {
-            error.should.not.exist();
-        }).then(() => {
-            accounts_nock.isDone().should.equal(true);
-        });
-    });
-
-    it('should succeed with cached access token info', () => {
-        let accounts_nock = nock(configuration.base_url)
-            .get(configuration.path + '/oauth/tokeninfo')
-            .query({access_token: 'jrrtoken'})
-            .reply(200, {scopes: scopes.join(' ')});
-
-        const request = httpMocks.createRequest({
-            method: 'GET',
-            url: '/test',
-            headers: {'Access-Token': 'jrrtoken'}
-        });
-
-
-        return accounts.verify_scopes(scopes)(
-            request,
-            null,
-            error => {
-                return error ? Promise.reject(error) : Promise.resolve();
-            }
-        ).catch(error => {
-            error.should.not.exist();
-        }).then(() => {
-            accounts_nock.isDone().should.not.equal(true);
-        });
+            });
     });
 });
